create tests and write readme.md and add scripts and add responses entity in controllers

+++ src/main/java/com/vlad/circledevs_test/models/DTO/AccountDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -11,5 +13,5 @@ public class AccountDTO {
   private Integer id;
   private String username;
   private String password;
-  private int cashAmount;
+  private BigDecimal cashAmount;
 }
